Ver1/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pip==23.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
